Add Windows installer
Currently it adds all of our JS deps so the installer installs a lot of files.
I'll optimize this alter with packaging the JS deps with an asar archive.

Also this installer has no default browser registry settings yet. I'll do a follow up for that as well.

Auditors: @KevinGrandon