Allow bookmarking and removing bookmarks
This makes the bookmark button in the navbar work.
It does not cover the bookmarks toolbar yet.

The implementation adds 2 new actions for removing and adding 'sites'.  This is a
generalized way of storing a list of sites.  Each site can be associated
with one or more tags, in paticular there is a 'bookmark' tag which
indicates that it is a bookmark.

Also refactored the functions for mapping key names to values to all use
a common function inside a new file funcitonal.js.  functional.js is
meant to be just for a collection of useful / common functional functions.

sieUtil.js is mostly an import from code I wrote in brave/browser with some minor
changes.

Auditors: @diracdeltas