Auto hide menu bar on Windows
On Windows it takes up too much space since we don't have a blended titlebar
there due to Electron limitations.  It's also the normal to only have the menu show
up when the Alt key is pressed on Windows. This auto hides it until Alt is
pressed like other applications on Windows.  I also noticed that you can use
camelCase for the titleBarStyle so I updated that too.

Auditors: @diracdeltas