Change titleBarStyle to help urlbar text selection
The issue is #118 but I'm leaving that open for now.

The problem isn't fixed yet but for whatever reason using hidden
instead of hidden-insert works a lot better for URLbar text selection.

There's some bug in electron that makes it so the region in the titlebar is draggable
which we'll have to fix eventually.

Auditors: @bridiver