Remove dependency on gnu rm for Windows builds
Windows doesn't have the gnu coreutils (including rm) by default, as such this script (which powers 'npm run build-package') cannot run successfully without the user first installing GnuWin or a similar tool bundle and wiring up the PATH environment variable to point to the location of the rm command.

This change removes the dependency on rm and uses Windows' own rmdir instead.

It also checks to see if any deletions are necessary by prefacing with 'if exist', and it eliminates all of the dist/* deletions, as they seemed unnecessary. (Perhaps they're required for other workflows I haven't got to yet, but on my Windows 10 machine, I never saw a dist folder created by the build.