re-add load-commit
I don't like calling loadStart and loadEnd twice but I need a work around for
the release today.  The changes for load-start are working a lot worse
because it is not called at all for the case of target=_blank.

Also if you navigate within a page like going to brianbondy.com and
clicking Block | Filters | Projects etc at the top, the location doesn't
change there either.

For target=blank, the urlbar url wasn't being set and no security icon etc.
Both functions seems idempotent to me though so this workaround works for now.

Auditors: @bridiver