More refactoring for siteutils
This code was getting to be unmanageable, this simplifies it even more.

The plan by the way is to add a 'key' for folders, similar to the
frameProps one and to distinguish based on that for folders.  Then each
site will have an optional parentFolderId and folders will have a
folderId plus an optional parentFolderId.

Before all this refactoring that plan would require changers all over.

Auditors: @diracdeltas