Bookmark management
This is a big chunk of work.  It adds a new context menu component which can be used in place of the electron one.  It has drag & drop support unlike electron's.  It will also allow us to add custom components inside menus later.

This also adds support for drag and drop within the bookmark manager and
within, into and out of folders on the bookmarks toolbar.

There's still a couple edge case things I need to handle but landing
this now.

Auditors: @diracdeltas