Only increment 307 redirect counter for frames
Sites like wired.com will send many requests for resources like tracking pixels
in a short period of time. This tricks the HTTPS Everywhere redirect heuristic
into thinking we are in an infinite redirect loop. Avoid this by only applying
the 307-counter heuristic on frame-type resources, which are the ones that tend
to be most noticeable by the user anyway.

Auditors: @bbondy