Implement spell check
This implements highlighting of misspelled words.  It also implements a context menu when right clicking on a misspelled word for up to the top 3 suggestions. You can also learn the word, in which case it will be suggested in future similar lookups, or ignore it so it won't be included in future lookups but will not be underlined

Fix #859

Auditors: @diracdeltas