Fix extension ID for PDFJS
Fix #2529

This was failing to load for me I think because the PDFJS_ORIGIN ID was different from the one installed. It's unclear to me if the new one that I have that is in use by my instance will be the same for everyone.  This works concistently for me though and keeps the 2 IDs in sync

Auditors: @bridiver