Address bar is not "active" when it should be
Makes the url bar active when switching to a about:newtab
page, as the user is expecting to be able to enter his url
when he switches back to a new tab page.

Fixes #3712

Auditors:
@bbondy

Test Plan:
1- Open a new tab
2- Switch to the new blank tab
3- Go back to the tab you were browsing before you opened the new tab
4- Go to the blank new tab you opened in step 1

On step 4, the address bar should be "active", so users could be able to write the web address without need to click on it.