Fix Redhat desktop launcher
The problem is that under RHEL the generated desktop launcher does not
have the correct mine types, so Brave can not be set as default browser.

This was reported and fixed here https://github.com/brave/browser-laptop/issues/2839
in both Brave and electron-redhat-installer package. However the fixed version
of the installer package was never used by Brave. This PR bumps the version
so it should work ok now.

Fixes #3776

Auditors:
@bbondy

Test Plan:
1. Install Brave under Fedora 24 and make sure it can be set as default
browser from the Default applications menu