Fix tabHandles to properly detect about pages
It did not have the ability to find about pages that had a hash at the end like you get when you switch to different sections in preferences.

I also adjusted the timeouts just because ledger APIs can take a long time, but that's not the main fix here.

Auditors: @bsclifton