Fix width of input element on New Bookmark Folder form
Fix #5192

Both input elements on the New Bookmark Folder form have the following
CSS properties set:

 * width: 100%;
 * max-width: 250px;
 * padding: 0 5px;

The default value for box-sizing is content-box (see
https://developer.mozilla.org/en-US/docs/Web/CSS/box-sizing). This
causes the input field for the bookmark folder's title to be 100% +
5px + 5px (width + padding left + padding right) wide. However, the
padding is not applied to the select field for parent folder selection,
causing the field to be only 100% wide (without the padding).

Using border-box as value for the box-sizing attribute for both form
fields renders them at an equal width.