tabIndex fixes
The call to setTabValues didn't do anything because it's stored outside of the dictionary in its own index.

The change in frame.js is actually because the prop on mount gets updated but the callback is called async so it ends up not getting new tabIndexes.  I'd imagine we'll eventually refactor this, but it makes it the same as setActive for now

Auditors: @bridiver