Remove tests for unsupported new-window code
There is no longer an event for new-window on webview. Instead this logic is handled with a process level add-new-contents event. There are already tests for these as well in the application window test.  I don't think they are fully exhaustive yet, but neither is what is being removed here.

Auditors: @bsclifton