Fix some pinned tab tests failing because of duplicate set-pinned messages
This no longer sends the windowId from tests for pinned and find
message.  It is assumed anyway that these tests should be called from
within the context of the specified window.

There's a bug where if you specify the windowId the app dispatcher from the main process will forward the message to all registered callbacks regardless of where they are registered from

We could solve this later if we need support by storing which windowId each renderer was registered from so that we can avoid calling into it, but for now that's not needed

Auditors: @bridiver