Disable minify on css
This was only a problem in production.

Minificaiton in CSS was causing problems with CSS variables and `getComputedStyle(document.querySelector(':root')).getPropertyValue('--bookmark-item-max-width')` returning non pixel values, which JS code expected it to be pixel values.
In particular it would only do this for any value that had more than 1
digit. It tried to use different unit types because it is less
characters.  This would be fine if there were no css variables being
used in JS.

Auditors: @bsclifton

Fix #6767