Re-enable ping to stats server for linux
This change re-enables the update check, used to send anonymous stats, for Linux.

The update check does not show the update bar, even if there is a newer version
of the software available. This will be addressed in a later commit.

Testing:

Testing this change from a QA perspective is problematic. These are the steps
I used to test the functionality of the change.

A)

  * Change the version in package.json to a previous release version
  * Build the binary on Linux
  * Launch the browser ensuring:
    * The browser does not crash on update init
    * The updater scheduler is initiated
    * Metadata is retrieved from the updater
    * The updater bar is NOT shown
    * The browser does not crash
    * No attempt is made to update the browser binary

B)

  * Change the browser version in package.json to a non-existent version
  * Build on Linux and launch
  * Ensure no metadata is retrieved
  * Ensure that the ping is recorded in the dw.fc_usage table of stats

Auditors: @bbondy

Implements: #7562