Update the correct frame's text on nav aborted
Auditors: @bsclifton

Fix #8251

If a background tab event came in it would cause the active frame's input text to change.  This could have caused some intermittent strangeness in urlbar previously as well

The fix is to pass the frame path to update to updateNavBarInput so it
doesn't use the active frame.