Fix broken torrent file view (fix for PR #8325)
Once a torrent is started, clicking on a video file within the torrent
should play it. Instead, the starting page is shown again and the user
has to restart the torrent. Once doing so, the file doesn't play. The
file list is still shown.

This is due to the ix= parameter not getting processed correctly on
hashchange after this change. Before, the redirect-based implementation
meant we didn't need to listen to hashchange since a redirect always
happened. Now, a file click actually just adds a hash to the URL and
triggers the hashchange event.