Add ability to update to a preview release
  * Add advanced preference to update to preview releases
  * Add parameter sent to the updater service to allow preview releases
  * Re-factor method to retrieve version for the current browser

Auditors: @bbondy, @clifton

Test plan:

This pull request has been extensively tested with a manual
parallel update system. It has been tested on both windows and
mac with the following scenarios:

  a) preview off - confirmed update to subsequent release version
  b) preview off - confirmed no update to preview version
  c) preview on - confirmed update to subsequent preview version

This change set is dependent on having the most current version
of the updater, with preview support, installed.

Relates to #8769